// packages/app/src/documents/types.ts
export interface DocumentWithOperationId {
  document: Document;
  postingOperationId: string | null;
  allowedActions: DocumentAction[];
}